+++ src/takajopkg/automagic.nim
@@ -40,11 +40,17 @@ proc autoMagic(level: string = "informational", skipProgressBar: bool = false,
   let cmd5 = ListIpAddressesCmd(name: "list-ip-addresses",
     displayTable: displayTable, timeline: timeline, output: output & "/ListIP-Addresses.txt")
 
-  # metrics-computers -t ../hayabusa/timeline.jsonl -o case-1/metricsComputers.csv
+  # metrics-computers -t ../hayabusa/timeline.jsonl -o case-1/MetricsComputers.csv
   let cmd0 = metricsComputersCmd(name: "metrics-computers",
     skipProgressBar: skipProgressBar,
     displayTable: displayTable,
-    timeline: timeline, output: output & "/metricsComputers.csv")
+    timeline: timeline, output: output & "/MetricsComputers.csv")
+
+  # metrics-users -t ../hayabusa/timeline.jsonl -o case-1/MetricsUsers.csv
+  let cmd01 = metricsUsersCmd(name: "metrics-users",
+    skipProgressBar: skipProgressBar,
+    displayTable: displayTable,
+    timeline: timeline, output: output & "/MetricsUsers.csv")
 
   # stack-cmdlines -t ../hayabusa/timeline.jsonl --level <leve> -o case-1/cmdlines.csv
   let cmd6 = StackCmdlineCmd(name: "stack-cmdlines", level: level,
@@ -146,7 +152,7 @@ proc autoMagic(level: string = "informational", skipProgressBar: bool = false,
     displayTable: displayTable, timeline: timeline, output: output & "/MitreTTP-Heatmap.json")
 
   # execute all command
-  let cmds = @[cmd1, cmd2, cmd3, cmd4, cmd5, cmd0, cmd6, cmd7, cmd8, cmd9, cmd10,
+  let cmds = @[cmd1, cmd2, cmd3, cmd4, cmd5, cmd0, cmd01, cmd6, cmd7, cmd8, cmd9, cmd10,
         cmd11, cmd12, cmd122, cmd13, cmd14, cmd15, cmd16, cmd17, cmd18, cmd19,
         cmd20,
         cmd21, cmd22, cmd23, cmd24, cmd25]